+++ SourceryRuntime/Sources/ProtocolComposition.swift
@@ -4,7 +4,7 @@
 import Foundation
 
 // sourcery: skipJSExport
-/// :nodoc:
+/// Describes a Swift [protocol composition](https://docs.swift.org/swift-book/ReferenceManual/Types.html#ID454).
 @objcMembers public final class ProtocolComposition: Type {
 
     /// Returns "protocolComposition"
@@ -14,6 +14,7 @@ import Foundation
     public let composedTypeNames: [TypeName]
 
     // sourcery: skipEquality, skipDescription
+    /// The types composed to form this composition, if known
     public var composedTypes: [Type]?
 
     /// :nodoc: