+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Sourcery CHANGELOG
 
 ---
+## Master
+
+
 ## 0.18.0
 
 ### New Features