+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ---
 ## Master
 
+### Bug fixes
+
+- Disallow protocol compositions from being considered as the `rawType` of an `enum` (#830)
 
 ## 0.18.0