+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Bug fixes
 
 - Disallow protocol compositions from being considered as the `rawType` of an `enum` (#830)
+- Add missing documentation for the `ProtocolComposition` type.
 
 ## 0.18.0