+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 - Disallow protocol compositions from being considered as the `rawType` of an `enum` (#830)
 - Add missing documentation for the `ProtocolComposition` type.
+- Fix incorrectly taking closure optional return value as sign that whole variable is optional (#823) 
 
 ## 0.18.0