+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Disallow protocol compositions from being considered as the `rawType` of an `enum` (#830)
 - Add missing documentation for the `ProtocolComposition` type.
 - Fix incorrectly taking closure optional return value as sign that whole variable is optional (#823) 
+- Fix incorrectly taking return values with closure as generic type as sign that whole variable is a closure (#845)
 
 ## 0.18.0