+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Sourcery CHANGELOG
 
 ---
-## Master
+## 1.0.0
 
 ### New Features